fix: 각 view의 response에서 pk(id)가 null로 표시되는 오류 수정

+++ otoeic/exam/models.py
@@ -8,7 +8,7 @@ from word.models import WordDAO
 
 
 class ExamDAO(models.Model):
-    id = models.IntegerField(primary_key=True, auto_created=True)
+    id = models.BigAutoField(primary_key=True, auto_created=True)
     user = models.ForeignKey(UserDAO, on_delete=models.CASCADE)
     level = models.IntegerField(choices=Level.choices)
     amount = models.IntegerField(validators=[MinValueValidator(10), MaxValueValidator(40)])
